Extract AdminGuardInner and skeleton, use Suspense
Refactor AdminGuard by extracting the client-side auth/redirect logic into a new AdminGuardInner component and moving the placeholder UI into AdminGuardSkeleton. The top-level AdminGuard now uses React.Suspense with AdminGuardSkeleton as the fallback and renders AdminGuardInner as the suspended child. This reduces duplication, separates concerns (UI skeleton vs auth logic), and keeps redirect behavior (unauthenticated -> login with ?next, non-admin -> user feed) intact.
